+++ Main.py
@@ -1,0 +1,17 @@
+import list_class,chemical_class
+
+if __name__ == "__main__":
+
+    chem_1 = Chemical('Chemical 1', 'CF1', 'Company1', 600, 500, 100)
+    chem_2 = Chemical('Chemical 2', 'CF2', 'Company2', 500, 400, 200)
+    chem_3 = Chemical('Chemical 3', 'CF3', 'Company3', 400, 300, 300)
+    chem_4 = Chemical('Chemical 4', 'CF4', 'Company4', 300, 200, 400)
+    chem_5 = Chemical('Chemical 5', 'CF5', 'Company5', 200, 100, 500)
+    chem_6 = Chemical('Chemical 6', 'CF6', 'Company6', 100, 50, 600)
+
+
+
+
+
+    chem_1.file_out()
+    chem_2.file_out()
