+++ list_class.py
@@ -1,0 +1,32 @@
+def node(self, prevnode=None, nextnode=None):
+    self.prev = prevnode
+    self.curr = self
+    self.next = nextnode
+
+class node:
+
+    def __init__(self,chemical):
+        chemical = chemical
+
+def displaylist(head):
+    currnode = head
+
+    while currnode is not None:
+        # we can then get this to display on the GUI
+        print(currnode.name)
+        currnode = currnode.next
+
+def addnode(chem)
+    temp = chem
+    if #something in list
+        while chem is not None
+
+    else:
+        chem.node(None,None)
+
+chem_1.node(None,chem_2)
+chem_2.node(chem_2,chem_3)
+chem_3.node(chem_3,chem_4)
+chem_4.node(chem_4,chem_5)
+chem_5.node(chem_5,chem_6)
+chem_6.node(chem_6)
